refactor: templates/manifest.tsをSSOTとして新設し、CLI/docsのテンプレート二重定義を解消
- templates/manifest.ts を新設し、Stack/CategoryId/TemplateDef 型と TEMPLATES 配列（7件）を一元管理
- templates/tsconfig.json を新設（eslint project service 用）
- lism-cli の create.ts から TEMPLATES 配列と型定義を削除し、manifest から import に変更
- lism-cli/tsconfig.json の include に manifest.ts を追加
- docs/src/config/templates.ts を manifest + docsMeta マージ方式に変更（title/thumb/draft は docs 専用メタで保持）
- docs/tsconfig.json の paths に @templates エイリアスを追加
- turbo.json の globalDependencies に manifest.ts を追加（manifest 変更で全タスク再実行）
