+++ apps/docs/tsconfig.json
@@ -5,7 +5,8 @@
   "compilerOptions": {
     "paths": {
       "@/*": ["./src/*"],
-      "@parts/*": ["./src/components/parts/*"]
+      "@parts/*": ["./src/components/parts/*"],
+      "@templates/*": ["../../templates/*"]
     }
   }
 }
