+++ packages/lism-cli/tsconfig.json
@@ -10,5 +10,5 @@
     "skipLibCheck": true,
     "types": ["node"]
   },
-  "include": ["src", "tsup.config.ts"]
+  "include": ["src", "tsup.config.ts", "../../templates/manifest.ts"]
 }
