+++ turbo.json
@@ -1,5 +1,6 @@
 {
   "$schema": "https://turbo.build/schema.json",
+  "globalDependencies": ["templates/manifest.ts"],
   "tasks": {
     "build": {
       "dependsOn": ["^build"],
